% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.r
\name{getters}
\alias{getters}
\alias{alias_expr}
\alias{alias_env}
\title{Query alias internals}
\usage{
alias_expr(alias, envir = parent.frame())

alias_env(alias, envir = parent.frame())
}
\arguments{
\item{alias}{the name of an alias, as a string}

\item{envir}{the environment in which to look up the alias name (default: calling environment)}
}
\value{
\code{alias_expr(alias)} returns an unevaluated R expression (a name or a call).

\code{alias_env(alias)} returns an environment.
}
\description{
\code{alias_expr(alias)} returns the expression that was used to define an alias.

\code{alias_env(alias)} returns the environment in which the aliased expression is evaluated.
}
